wip: restructure app
