+++ compose/neurosynth_compose/tests/api/test_meta_analysis_result.py
@@ -1,6 +1,8 @@
 from pathlib import Path
 
 from ...models import MetaAnalysis, User
+from ...resources.tasks import file_upload_neurovault
+
 
 def test_meta_analysis_result(app, auth_client, user_data, meta_analysis_results, mock_pynv):
     user = User.query.filter_by(name="user1").first()
@@ -19,3 +21,8 @@ def test_meta_analysis_result(app, auth_client, user_data, meta_analysis_results
     data["meta_analysis_id"] = meta_analysis.id
     post_result = auth_client.post("/api/meta-analysis-results", data=data)
     assert post_result.status_code == 200
+
+
+def test_file_upload_neurovault(app, user_data):
+
+    file_upload_neurovault('tat', 'bt')
